third_party/blink/web_tests/platform/mac-mac15-arm64/external/wpt/webxr/idlharness.https.window-expected.txt

This is a testharness.js-based test.
Found 29 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] idl_test setup
  promise_test: Unhandled rejection with value: object "TypeError: Failed to construct 'XRWebGLLayer': The provided value is not of type '(WebGL2RenderingContext or WebGLRenderingContext)'."
[FAIL] WebGLRenderingContext includes WebGLRenderingContextOverloads: member names are unique
  assert_true: member bufferData is unique expected true got false
[FAIL] XRSession interface: attribute trackedSources
  assert_true: The prototype object must have a property "trackedSources" expected true got false
[FAIL] XRSession interface: xrSession must inherit property "trackedSources" with the proper type
  assert_inherits: property "trackedSources" not found in prototype chain
[FAIL] XRRenderState interface: attribute passthroughFullyObscured
  assert_true: The prototype object must have a property "passthroughFullyObscured" expected true got false
[FAIL] XRRenderState interface: xrRenderState must inherit property "passthroughFullyObscured" with the proper type
  assert_inherits: property "passthroughFullyObscured" not found in prototype chain
[FAIL] XRFrame interface: attribute predictedDisplayTime
  assert_true: The prototype object must have a property "predictedDisplayTime" expected true got false
[FAIL] XRInputSource interface: attribute skipRendering
  assert_true: The prototype object must have a property "skipRendering" expected true got false
[FAIL] XRWebGLLayer interface: attribute fixedFoveation
  assert_true: The prototype object must have a property "fixedFoveation" expected true got false
[FAIL] XRWebGLLayer must be primary interface of xrWebGLLayer
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: xrWebGLLayer is not defined"
[FAIL] Stringification of xrWebGLLayer
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: xrWebGLLayer is not defined"
[FAIL] XRWebGLLayer interface: xrWebGLLayer must inherit property "antialias" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: xrWebGLLayer is not defined"
[FAIL] XRWebGLLayer interface: xrWebGLLayer must inherit property "ignoreDepthValues" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: xrWebGLLayer is not defined"
[FAIL] XRWebGLLayer interface: xrWebGLLayer must inherit property "fixedFoveation" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: xrWebGLLayer is not defined"
[FAIL] XRWebGLLayer interface: xrWebGLLayer must inherit property "framebuffer" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: xrWebGLLayer is not defined"
[FAIL] XRWebGLLayer interface: xrWebGLLayer must inherit property "framebufferWidth" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: xrWebGLLayer is not defined"
[FAIL] XRWebGLLayer interface: xrWebGLLayer must inherit property "framebufferHeight" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: xrWebGLLayer is not defined"
[FAIL] XRWebGLLayer interface: xrWebGLLayer must inherit property "getViewport(XRView)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: xrWebGLLayer is not defined"
[FAIL] XRWebGLLayer interface: calling getViewport(XRView) on xrWebGLLayer with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: xrWebGLLayer is not defined"
[FAIL] XRWebGLLayer interface: xrWebGLLayer must inherit property "getNativeFramebufferScaleFactor(XRSession)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: xrWebGLLayer is not defined"
[FAIL] XRWebGLLayer interface: calling getNativeFramebufferScaleFactor(XRSession) on xrWebGLLayer with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: xrWebGLLayer is not defined"
[FAIL] XRPermissionStatus interface: existence and properties of interface object
  assert_own_property: self does not have own property "XRPermissionStatus" expected property "XRPermissionStatus" missing
[FAIL] XRPermissionStatus interface object length
  assert_own_property: self does not have own property "XRPermissionStatus" expected property "XRPermissionStatus" missing
[FAIL] XRPermissionStatus interface object name
  assert_own_property: self does not have own property "XRPermissionStatus" expected property "XRPermissionStatus" missing
[FAIL] XRPermissionStatus interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "XRPermissionStatus" expected property "XRPermissionStatus" missing
[FAIL] XRPermissionStatus interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "XRPermissionStatus" expected property "XRPermissionStatus" missing
[FAIL] XRPermissionStatus interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "XRPermissionStatus" expected property "XRPermissionStatus" missing
[FAIL] XRPermissionStatus interface: attribute granted
  assert_own_property: self does not have own property "XRPermissionStatus" expected property "XRPermissionStatus" missing
[FAIL] WebGLRenderingContextBase interface: webGLRenderingContextBase must inherit property "makeXRCompatible()" with the proper type
  assert_inherits: provided value is not an object
Harness: the test ran to completion.
